(MCO-739) Fix negative data plugin comparisons
In the past when data plugins lookups resolved to false, it was
impossible for us to do a false comparison.

The statement:

```
puppet().enabled=false
```

Would always evaluate to false, even when the statement was translated
into

```
false == false
```

Here we fix the bug and do some light refactoring to #eval_compound_fstatement
while we're there.