(MCO-742) Fix MCollective::RPC#run on Windows
systemu needs to be able to figure out a path to a working Ruby
executable in order to work correctly. The systemu defaults previously
didn't reliably work on Windows due to possible spaces in the path to
the executable.

This patch to the vendored systemu allows full paths to executables to
have spaces by switching to individual argument invocation of system(),
instead of relying on system() to correctly parse a single string into
executable + arguments.

There are two main ways to invoke system() to execute a command and
return an exit code.

1. Pass a single string argument which will be evaluated as if the
string was typed into a shell. That is, the system will try to correctly
identify and split the string into an executable path followed by
arguments to the executable. The call system("/bin/echo one two") would
be broken down to an argument vector of ["/bin/echo", "one", "two"]
because Ruby would split it on spaces.

2. Pass an explicit argument vector to system() where each argument
passed is an element of the argument vector. In this invocation, rather
than passing a single string, multiple arguments are passed thusly:
system("/bin/echo", "one", "two"). Doesn't matter whether or not there
are spaces in any of these, Ruby isn't parsing them and will use the
argument vector exactly as given.

The explicit argv invocation is preferable because just like when typing
a command into a shell, if the path to the executable contains spaces,
the implicit invocation will incorrectly split on the space in the path
to the executable, and incorrectly try to execute at best a non-existent
file or directory. Consider:

system("C:\Program Files\Puppet\bin\puppet agent -t")

This call would fail because the argument vector would be built out as:
["C:\Program", "Files\Puppet\bin\puppet", "agent", "-t"]

Better would be to instead use
system(""C:\Program Files\Puppet\bin\puppet", "agent", "-t").

This is effectively what this commit does.

The "ruby" invocation is changed also but only for consistency.