(MCO-742) Fix MCollective::RPC#run on Windows
systemu needs to be able to figure out a path to a working Ruby
executable in order to work correctly. The systemu defaults previously
didn't reliably work on Windows due to possible spaces in the path to
the executable.

This patch to the vendored systemu allows full paths to executables to
have spaces by switching to individual argument invocation of system(),
instead of relying on system() to correctly parse a single string into
executable + arguments.