(MCO-745) Add windows support to install.rb
We need to be able to support installing Marionette Collective with the
provided install.rb file on windows. This commit does that.

We need to avoid running File.join on our paths on windows, because it
likes to add a `/` to the beginning of the path, so we end up with
`/C:/`, which throws an error.