(MCO-748) Update mco.bat to reflect current standard
With the switch over to building puppet-agent for windows with Vanagon,
we need to effectively install mcollective with install.rb. This script
installs mco.bat, which is a wrapper that lets windows handle ruby
scripts. This commit updates mco.bat to use our environment.bat file if
we install it, and to be more in line with the mco.bat file installed
with the puppet-agent package as built by puppet_for_the_win

This commit also updates install.rb to not add `#!` to find ruby on
windows. This syntax is not understood on windows.