(MCO-762) Add --no-batch-files option
puppet-agent carries the batch files for all executables with the new move to
vanagon for windows projects, we will need an option to omit batch file
creation during installation of mco so there are not more than one set of
batch files. So this commit adds a --no-batch-files option for cmake that omits
those files.