(MCO-765) don't use eval un unchecked arguments
This patch drops the use of `eval` to evaluate comparison expressions in
discovery filters. Instead the corresponding comparison method is called
on the arguments directly while coercing the arguments to appropriate data
types.
While this approach might produce different results than the original
implementation for expressions which relied on the `eval` to perform more
than just the comparison, it closes a whole lot of vulnerabilities stemming
from insufficient checking of the comparison arguments.