(maint) proper handling of quoted strings
Before this commit a mere presence of a quoted string in a discovery filter
expression caused possible function call in the same expression to be treated
as a literal argument rather than a function call.
This improves the parsing of the quoted strings in the expressions to not have
that side effect.