(MCO-775) Include discovery timeout when making requests.
The discovery timeout sets a base level for how long it takes agents to
respone to requests. It should be included along with any filters when
computing the discovrey timeout for all requests. Before this commit
only the --timeout or the timeout of specific agents was used. Now the
discovery timeout is added to that time to account for system wide
performance.