(MCO-778) use publish_timeout from config
The Mcollective client class was previously ignoring the publish_timeout
value when specified in the config file, and by default would wait
forever when unthreaded and would crash when threaded. This commit fixes
this by preferring an explicitly set `publish_timeout` value and then
falling back to the config set value.