(maint) Do not build debs/rpms
Since we've integrated marionette collective into the puppet-agent
package, we do not want to build deb or rpm packages. We only want the
gem and tarball to be created when we kick off our build automation.
However, as there are some community members that use this automation,
we should keep the relics in this file for future reference. That is why
they are only commented out, and not fully removed.