(MCO-785) Fix regression in --batch option
It looks like changes introduced in 928ccac63ae2820bc47679264c12e4adf3ba08c6
broke --batch option, because one of stats hash elements is not
initialized as Array and concat function is called on that object which
leads to crash.

This commit will make sure, that unexpectedresponsefrom key is
initialized as Array.