(MCO-799) Consider string and symbol keys equivelant
This has the end effect that:

   c.get_fact(:fact => "cluster")

is the same as:

   c.get_fact("fact" => "cluster")

in the case where the DDL defined the input :fact for the action
:get_fact

Additionally RPC::Request will now intelligently convert request[:foo]
into fetching "foo" from such a request.  We could perhaps munge the
entire incoming request instead of this approach but this seems simpler
to implement

Long term I want to downgrade MCollective network protocols to be JSON
pure to facilitate non Ruby clients and JSON based REST gateways and
just general removal of user supplied YAML.

This changes makes it so that not everyone have to redevelop their DDLs,
Agents and Clients

The one problem case is where someone legit made a input :foo and also
one "foo" on the same action, if all we have is a string to represent
the input - "foo" - we cannot distinguish between them, such agents will
now warn on the client.  I anticipate this to be extreme edge cases as
we've been really good at recommending symbols in all docs and all our
agents