(MCO-816) Allow upgrading puppet-agent with mcollective
Since mcollective is part of puppet-agent packaging, upgrading that
package over an mcollective plugin is tricky. Previously the mcollective
systemd service was currently configured in a way that broke that,
because it killed the entire process tree when stopping as part of the
upgrade.

Set `KillMode=process` for the systemd service (similar to how Puppet
and PXP Agent are configured) so that it can be restarted during the
upgrade without killing the upgrade itself.