(MCO-803) Log unexpected STOMP frames at WARN level
We raise an error message if a STOMP frame with a type other than MESSAGE is
received. Details about the frame are logged at debug level, which means they
never actually make it into the logs on a normal system. The content of these
frames can include important information about broker health. This commit
increases the log level from DEBUG to WARN so that this information is captured
alongside the error message.